+++ src/NexusMods.DataModel/Abstractions/EntityLink.cs
@@ -1,6 +1,6 @@
-using System.Text.Json;
 using System.Text.Json.Serialization;
 using NexusMods.DataModel.Abstractions.Ids;
+using NexusMods.DataModel.JsonConverters;
 
 namespace NexusMods.DataModel.Abstractions;
 
@@ -60,53 +60,4 @@ public record struct EntityLink<T> : IEmptyWithDataStore<EntityLink<T>>,
 
     /// <summary/>
     public static implicit operator EntityLink<T>(T t) => new(t.DataStoreId, t.Store);
-}
-
-/// <inheritdoc />
-public class EntityLinkConverterFactory : JsonConverterFactory
-{
-    private readonly IServiceProvider _provider;
-
-    /// <inheritdoc />
-    public EntityLinkConverterFactory(IServiceProvider provider)
-    {
-        _provider = provider;
-    }
-
-    /// <inheritdoc />
-    public override bool CanConvert(Type typeToConvert) =>
-        typeToConvert.GenericTypeArguments.Length == 1 &&
-        typeToConvert.GetGenericTypeDefinition() == typeof(EntityLink<>);
-
-    /// <inheritdoc />
-    public override JsonConverter CreateConverter(Type typeToConvert, JsonSerializerOptions options)
-    {
-        return (JsonConverter)_provider.GetService(typeof(EntityLinkConverter<>).MakeGenericType(typeToConvert.GenericTypeArguments))!;
-    }
-}
-
-/// <inheritdoc />
-public class EntityLinkConverter<T> : JsonConverter<EntityLink<T>>
-    where T : Entity
-{
-    private readonly IDataStore _store;
-
-    /// <inheritdoc />
-    public EntityLinkConverter(IDataStore store)
-    {
-        _store = store;
-    }
-
-    /// <inheritdoc />
-    public override EntityLink<T> Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-    {
-        var id = JsonSerializer.Deserialize<IId>(ref reader, options)!;
-        return new EntityLink<T>(id, _store);
-    }
-
-    /// <inheritdoc />
-    public override void Write(Utf8JsonWriter writer, EntityLink<T> value, JsonSerializerOptions options)
-    {
-        JsonSerializer.Serialize(writer, value.Id, options);
-    }
 }
