+++ src/NexusMods.DataModel/Abstractions/Ids/IId.cs
@@ -1,6 +1,6 @@
 using System.Buffers.Binary;
-using System.Text.Json;
 using System.Text.Json.Serialization;
+using NexusMods.DataModel.JsonConverters;
 
 namespace NexusMods.DataModel.Abstractions.Ids;
 
@@ -149,28 +149,5 @@ public interface IId
         Span<byte> span = stackalloc byte[16];
         guid.TryWriteBytes(span);
         return new IdVariableLength(category, span.ToArray());
-    }
-}
-
-/// <inheritdoc />
-public class IdJsonConverter : JsonConverter<IId>
-{
-    /// <inheritdoc />
-    public override IId Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-    {
-        var str = reader.GetString()!;
-        var spanSize = (int)Math.Ceiling((double)str.Length / 4) * 3;
-        Span<byte> span = stackalloc byte[spanSize];
-        Convert.TryFromBase64String(str, span, out _);
-
-        return IId.FromTaggedSpan(span);
-    }
-
-    /// <inheritdoc />
-    public override void Write(Utf8JsonWriter writer, IId value, JsonSerializerOptions options)
-    {
-        Span<byte> span = stackalloc byte[value.SpanSize + 1];
-        value.ToTaggedSpan(span);
-        writer.WriteBase64StringValue(span);
     }
 }
