+++ src/NexusMods.DataModel/Loadouts/ModFileId.cs
@@ -1,5 +1,5 @@
-using System.Text.Json;
 using System.Text.Json.Serialization;
+using NexusMods.DataModel.JsonConverters;
 using Vogen;
 
 namespace NexusMods.DataModel.Loadouts;
@@ -11,21 +11,5 @@ public partial struct ModFileId
     public static ModFileId New()
     {
         return From(Guid.NewGuid());
-    }
-
-    public class ModFileIdConverter : JsonConverter<ModFileId>
-    {
-        public override ModFileId Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-        {
-            var data = reader.GetBytesFromBase64();
-            return From(new Guid(data));
-        }
-
-        public override void Write(Utf8JsonWriter writer, ModFileId value, JsonSerializerOptions options)
-        {
-            Span<byte> span = stackalloc byte[16];
-            value._value.TryWriteBytes(span);
-            writer.WriteBase64StringValue(span);
-        }
     }
 }
