+++ src/NexusMods.DataModel/Loadouts/ModId.cs
@@ -1,5 +1,5 @@
-using System.Text.Json;
 using System.Text.Json.Serialization;
+using NexusMods.DataModel.JsonConverters;
 using Vogen;
 
 namespace NexusMods.DataModel.Loadouts;
@@ -12,21 +12,4 @@ public partial struct ModId
     {
         return From(Guid.NewGuid());
     }
-
-    public class ModIdConverter : JsonConverter<ModId>
-    {
-        public override ModId Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-        {
-            var data = reader.GetBytesFromBase64();
-            return From(new Guid(data));
-        }
-
-        public override void Write(Utf8JsonWriter writer, ModId value, JsonSerializerOptions options)
-        {
-            Span<byte> span = stackalloc byte[16];
-            value._value.TryWriteBytes(span);
-            writer.WriteBase64StringValue(span);
-        }
-    }
 }
-
